[#44] Moved get_ref_vector() to _class1_uniformize.py + started uniformize(), TBF
